ai engine/requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
openai>=1.30.0
apscheduler>=3.10.0
httpx>=0.27.0
arxiv>=2.1.0
huggingface-hub>=0.25.0
tavily-python>=0.5.0
reportlab>=4.0.0
